requirements.txt
flask>=2.3.3
pandas>=2.2.0
gunicorn>=21.2.0
facebook-sdk==3.1.0